+++ src/onegov.election_day/onegov/election_day/assets/js/common.js
@@ -14,6 +14,17 @@ $('.collapsible .more, .collapsible .less').click(function() {
 // sort tables wishing to be sorted
 $('table.sortable').tablesorter({widgets: ['staticRow']});
 
+// Add backend dropdown actions
+$('ul.actions').each(function(index, element) {
+    $(element).before(
+        $('<a></a>')
+            .attr('href', '#')
+            .attr('class', 'button split small action-button')
+            .on('click', function() { $(element).toggle(); return false; })
+            .html($(this).data('title') + ' <span></span>')
+    ).hide();
+});
+
 // force all dropdowns to be rendered in the direction specified in the
 // options of said dropdown (so if we say align:left, *always* align left)
 Foundation.libs.dropdown.small = function() { return false; };